new mod header.rs, split from block.rs
